Rename the trans associate files
Rename the files dialog-trans-assoc.c/h and the glade file to the less
specific form dialog-assoc.c/h so that it can be used for other
associations.
