+++ gnucash/gnome/dialog-assoc.h
