+++ gnucash/gtkbuilder/dialog-assoc.glade
